ai_service/requirements.txt
fastapi>=0.115.0
python-jose[cryptography]>=3.3.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
openai>=1.50.0
openai-agents>=0.0.14
pinecone>=3.0.0
boto3>=1.35.0
httpx>=0.27.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
tiktoken>=0.7.0
